requirements.txt
streamlit>=1.21.0
altair<5
pytest
python-dotenv>=1.0.0
google-generativeai>=0.5.0
